// backend/src/modules/payment/enums/payment-mode.enum.ts
export enum PAYMENT_MODE {
  Cash = 'payment.payment_mode.cash',
  CreditCard = 'payment.payment_mode.credit_card',
  Check = 'payment.payment_mode.check',
  BillOfExchange = 'payment.payment_mode.bill_of_exchange',
  BankTransfer = 'payment.payment_mode.bank_transfer',
  WireTransfer = 'payment.payment_mode.wire_transfer',
  CreditNoteSettlement = 'payment.payment_mode.credit_note_settlement',
}