// backend/src/modules/document-template/enums/document-template-document-type.enum.ts
export enum DOCUMENT_TEMPLATE_DOCUMENT_TYPE {
  INVOICE = 'INVOICE',
  QUOTE = 'QUOTE',
  CUSTOMER_ORDER = 'CUSTOMER_ORDER',
  DELIVERY_NOTE = 'DELIVERY_NOTE',
  GOODS_ISSUE_NOTE = 'GOODS_ISSUE_NOTE',
  CREDIT_NOTE = 'CREDIT_NOTE',
  RETURN_NOTE = 'RETURN_NOTE',
  PURCHASE_ORDER = 'PURCHASE_ORDER',
  SUPPLIER_INVOICE = 'SUPPLIER_INVOICE',
  SUPPLIER_CREDIT_NOTE = 'SUPPLIER_CREDIT_NOTE',
  RECEIPT = 'RECEIPT',
  CUSTOM_DOCUMENT = 'CUSTOM_DOCUMENT',
}